self-parking-ai-2d/requirements.txt
neat-python==0.92
numpy~=1.21.1
pygame==2.0.1